src/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[project]
name = "seeact"
version = "0.2.9.0"
description = "SeeAct is a system for generalist web agents that autonomously carry out tasks on any given website, with a focus on large multimodal models (LMMs) such as GPT-4V(ision). It consists of two main components: (1) A robust codebase that supports running web agents on live websites, and (2) an innovative framework that utilizes LMMs as generalist web agents."
authors = [
    {name = "Boyu Gou", email = "gou.43@osu.edu"},
    {name = "Boyuan Zheng", email = "zheng.2372@osu.edu"},
    {name = "Zheng Du", email = "du.913@osu.edu"}
]
license = {file="LICENSE"}
readme = "https://github.com/OSU-NLP-Group/SeeAct/blob/main/README.md"
requires-python = ">=3.9"

dependencies = [
    "backoff",
    "playwright",
    "toml",
    "httpx==0.27.2",
    "litellm==1.35.32",
    "python-dotenv"
]

[project.urls]
homepage = "https://osu-nlp-group.github.io/SeeAct/"
repository = "https://github.com/OSU-NLP-Group/SeeAct"
documentation = "https://github.com/OSU-NLP-Group/SeeAct"
